+++ src/pages/Register/index.tsx
@@ -1,26 +1,13 @@
-import React, { FormEvent, useRef, useState } from 'react'
+import React, { FormEvent, useCallback, useRef, useState } from 'react'
 import {
   postRegister,
   postConfirmVerificationCode,
   postCreateVerificationCode,
 } from '../../utils/api'
-import {
-  chakra,
-  FormControl,
-  InputGroup,
-  InputLeftElement,
-  Stack,
-  Text,
-  Input,
-  Spinner,
-  Flex,
-  useToast,
-} from '@chakra-ui/react'
-import theme from '../../style/theme'
+import { chakra, Text, useToast } from '@chakra-ui/react'
 import { FaUserAlt, FaLock } from 'react-icons/fa'
 import {
   RegisterStack,
-  RegisterText,
   RegisterButton,
   ErrorText,
   DeleteBtn,
@@ -33,14 +20,10 @@ import { useDispatch, useSelector } from 'react-redux'
 import { RootState } from '../../redux/store'
 import { setWhichInfo } from '../../redux/register/registerSlice'
 import LoadingSpinner from '../../components/LoadingSpinner'
+import UserInput from '../../components/UserInput'
 
 const Register = (): JSX.Element => {
-  const CFaUserAlt = chakra(FaUserAlt)
-  const CFaLock = chakra(FaLock)
   const navigate = useNavigate()
-  const emailRef = useRef<HTMLInputElement>(null)
-  const verificationRef = useRef<HTMLInputElement>(null)
-  const passwordRef = useRef<HTMLInputElement>(null)
   const [input, setInput] = useState<RegisterForm>({
     username: '',
     password: '',
@@ -54,8 +37,7 @@ const Register = (): JSX.Element => {
   const whichUI = useSelector((state: RootState) => state.register.whichUI)
   const agreeInfo = useSelector((state: RootState) => state.register.agreeInfo)
   const [errorText, setErrorText] = useState('')
-
-  const handleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+  const handleChange = useCallback((e: React.ChangeEvent<HTMLInputElement>) => {
     const { name, value } = e.target
     if (name === 'verificationCode') {
       if (value.length > 6) {
@@ -65,7 +47,7 @@ const Register = (): JSX.Element => {
       return
     }
     setInput({ ...input, [name]: value })
-  }
+  }, [])
 
   const handleAgreeBtn = () => {
     if (agreeInfo.private === true && agreeInfo.service === true) {
@@ -90,7 +72,7 @@ const Register = (): JSX.Element => {
     setErrorText(message)
     if (message === '이미 존재하는 회원입니다.') {
       setInput({ ...input, username: '' })
-      emailRef.current && emailRef.current.focus()
+      // emailRef.current && emailRef.current.focus()
       dispatch(setWhichInfo('Loading'))
     } else if (status < 400) {
       dispatch(setWhichInfo('Loading'))
@@ -114,7 +96,7 @@ const Register = (): JSX.Element => {
     switch (message) {
       case '존재하지 않는 인증 정보입니다.': {
         setInput({ ...input, verificationCode: '' })
-        verificationRef.current && verificationRef.current.focus()
+        // verificationRef.current && verificationRef.current.focus()
         break
       }
       case '만료된 인증 코드입니다.': {
@@ -173,7 +155,7 @@ const Register = (): JSX.Element => {
         dispatch(setWhichInfo('Valid'))
       }
     } else {
-      passwordRef.current && passwordRef.current.focus()
+      // passwordRef.current && passwordRef.current.focus()
       setErrorText('비밀번호가 일치하지 않습니다')
     }
   }
@@ -213,83 +195,47 @@ const Register = (): JSX.Element => {
           <Title>회원가입</Title>
           {!whichUI.isAgreed && <TermsOfUse></TermsOfUse>}
           {whichUI.isAgreed && (
-            <FormControl isRequired>
-              <RegisterText>이메일</RegisterText>
-              <InputGroup borderColor={`${theme.color.ligthGray}`}>
-                <InputLeftElement>
-                  {<CFaUserAlt color="gray.300" />}
-                </InputLeftElement>
-                <Input
-                  ref={emailRef}
-                  type="email"
-                  name="username"
-                  placeholder="지스트 메일을 입력하세요"
-                  value={input.username}
-                  onChange={handleChange}
-                  disabled={whichUI.isCodeRequested}
-                  borderRadius="0"
-                ></Input>
-              </InputGroup>
-            </FormControl>
+            <UserInput
+              text="이메일"
+              name="username"
+              type="email"
+              value={input.username}
+              placeholder="지스트 메일을 입력하세요"
+              onChange={handleChange}
+              disabled={whichUI.isCodeRequested}
+            ></UserInput>
           )}
 
           {whichUI.isCodeRequested && !whichUI.isExpired && (
-            <FormControl isRequired>
-              <Text mb="8px">인증 코드</Text>
-              <InputGroup borderColor={`${theme.color.ligthGray}`}>
-                <InputLeftElement>
-                  {<CFaLock color="gray.300" />}
-                </InputLeftElement>
-                <Input
-                  ref={verificationRef}
-                  type="text"
-                  name="verificationCode"
-                  placeholder="이메일로 온 인증 코드를 입력하세요"
-                  value={input.verificationCode}
-                  onChange={handleChange}
-                  disabled={whichUI.isVerificated}
-                  style={{ textTransform: 'uppercase' }}
-                  borderRadius="0"
-                ></Input>
-              </InputGroup>
-            </FormControl>
+            <UserInput
+              text="인증 코드"
+              name="verificationCode"
+              type="text"
+              value={input.verificationCode}
+              placeholder="이메일로 온 인증 코드를 입력하세요"
+              onChange={handleChange}
+              disabled={whichUI.isVerificated}
+            ></UserInput>
           )}
           {whichUI.isVerificated && (
-            <FormControl isRequired>
-              <Text mb="8px">비밀번호</Text>
-              <InputGroup borderColor={`${theme.color.ligthGray}`}>
-                <InputLeftElement>
-                  {<CFaLock color="gray.300" />}
-                </InputLeftElement>
-                <Input
-                  ref={passwordRef}
-                  type="password"
-                  name="password"
-                  placeholder="영문과 숫자를 포함한 8자리 이상의 비밀번호를 입력하세요"
-                  value={input.password}
-                  onChange={handleChange}
-                  borderRadius="0"
-                ></Input>
-              </InputGroup>
-            </FormControl>
+            <UserInput
+              text="비밀번호"
+              name="password"
+              type="password"
+              value={input.password}
+              placeholder="영문과 숫자를 포함한 8자리 이상의 비밀번호를 입력하세요"
+              onChange={handleChange}
+            ></UserInput>
           )}
           {whichUI.isVerificated && (
-            <FormControl isRequired>
-              <Text mb="8px">비밀번호 확인</Text>
-              <InputGroup borderColor={`${theme.color.ligthGray}`}>
-                <InputLeftElement>
-                  {<CFaLock color="gray.300" />}
-                </InputLeftElement>
-                <Input
-                  type="password"
-                  name="passwordConfirm"
-                  placeholder="비밀번호를 재입력하세요"
-                  value={input.passwordConfirm}
-                  onChange={handleChange}
-                  borderRadius="0"
-                ></Input>
-              </InputGroup>
-            </FormControl>
+            <UserInput
+              text="비밀번호 확인"
+              name="passwordConfirm"
+              type="password"
+              value={input.passwordConfirm}
+              placeholder="비밀번호를 재입력하세요"
+              onChange={handleChange}
+            ></UserInput>
           )}
           {!whichUI.isAgreed && (
             <RegisterButton onClick={handleAgreeBtn}>다음단계</RegisterButton>
@@ -307,9 +253,7 @@ const Register = (): JSX.Element => {
               인증코드 재전송
             </RegisterButton>
           )}
-
           {whichUI.isLoading && <LoadingSpinner></LoadingSpinner>}
-
           {whichUI.isCodeRequested &&
             !whichUI.isVerificated &&
             !whichUI.isExpired && (
