+++ src/redux/register/registerSlice.ts
@@ -24,7 +24,6 @@ export const registerSlice = createSlice({
   initialState,
   reducers: {
     setWhichInfo: (state, action) => {
-      console.log(action.payload)
       switch (action.payload) {
         case 'Agreed':
           state.whichUI.isAgreed = !state.whichUI.isAgreed
